+++ projects/ngx-dmaster-ui/src/lib/components/data-display/table/index.ts
@@ -1,3 +1,5 @@
 export * from './table.component';
+export * from './table-cell.directive';
+export * from './table-empty.directive';
 export * from './table.tokens';
 export * from './table.types';